fixed to a float

+++ planets.py
@@ -1,5 +1,5 @@
 def weight_on_planets():
-   earthWeight = int(input("What do you weigh on earth? "))
+   earthWeight = float(input("What do you weigh on earth? "))
 
    print("\nOn Mars you would weigh %s pounds.\nOn Jupiter you would weigh %s pounds." % (earthWeight*0.38, earthWeight*2.34))
